backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["zhkuo24 <zhkuo1989@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.63.0"}
python-dotenv = "^0.15.0"
loguru = "^0.5.3"
SQLAlchemy = "^1.3.22"
python-jose = "^3.2.0"
bcrypt = "^3.2.0"
emails = "^0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
